+++ src/AutoDeconstruct/AutoDeconstructAttribute.cs
@@ -4,7 +4,7 @@
 /// Instructs AutoDeconstruct to generate a <c>Deconstruct()</c> implementation.
 /// </summary>
 [AttributeUsage(AttributeTargets.Assembly | AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Interface,
-	AllowMultiple = false, Inherited = false)]
+	AllowMultiple = true, Inherited = false)]
 public sealed class AutoDeconstructAttribute
 	: Attribute
 {