+++ changelog.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.0.0] - Not Yet Released
 
+### Added
+- `[AutoDeconstruct]` can now be used to target any type (issue [#19](https://github.com/JasonBock/AutoDeconstruct/issues/19))
+
 ### Changed
 - Removed the ability to add deconstruction to every time in an assembly (issue [#18](https://github.com/JasonBock/AutoDeconstruct/issues/18))